refactor: Renamed AetherRendering to AetherRenderers, and moved anything in it that wasn't related to a renderer into AetherClient. Also added annotations for some event subscribers instead of registering through the event bus methods.
