+++ src/main/java/com/gildedgames/aether/common/registry/AetherBlocks.java
@@ -3,10 +3,8 @@ package com.gildedgames.aether.common.registry;
 import java.util.Objects;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.function.ToIntFunction;
 
 import com.gildedgames.aether.Aether;
-import com.gildedgames.aether.client.AetherRendering;
 import com.gildedgames.aether.client.registry.AetherParticleTypes;
 import com.gildedgames.aether.common.block.construction.AerogelBlock;
 import com.gildedgames.aether.common.block.construction.AerogelSlabBlock;
@@ -44,16 +42,12 @@ import com.gildedgames.aether.common.block.utility.SunAltarBlock;
 import com.gildedgames.aether.common.item.block.BurnableBlockItem;
 import com.gildedgames.aether.common.world.gen.tree.GoldenOakTree;
 import com.gildedgames.aether.common.world.gen.tree.SkyrootTree;
-import com.google.common.collect.ImmutableMap;
 
-import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.material.Material;
 import net.minecraft.world.level.material.MaterialColor;
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.item.BedItem;
 import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.HoeItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Rarity;
 import net.minecraft.world.item.SignItem;
@@ -71,7 +65,6 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.DoorBlock;
 import net.minecraft.world.level.block.FenceBlock;
 import net.minecraft.world.level.block.FenceGateBlock;
-import net.minecraft.world.level.block.FireBlock;
 import net.minecraft.world.level.block.FlowerBlock;
 import net.minecraft.world.level.block.FlowerPotBlock;
 import net.minecraft.world.level.block.GlassBlock;
